Complete task-01 in homeworks/01
Task was completed by applying position relative to the .container
parent element, and position absolute to the .green and .yellow child
elements, along with changing their top and left properties.

Even though the .green and .yellow elements are not direct children of
.container, they are positioned relative to it because .container is
the closest parent element with a position different from static.

Since, by default, overlapping elements are displayed in the order
they are written in the HTML, the .yellow element is displayed under
the .green element. To fix this, the z-index property was set to 1 on
the .yellow element, as it allows changing the stacking order of
elements with position different from static.