Update header in homeworks/01 task-02
Added display flex and justify-content center on .header element to
make flex items appear next to each other and centered. Also, added
position sticky with top 0 to make the header dynamically change its
position to fixed and stay at the top of the viewport when the page
is scrolled. With media query, the header is hidden when the screen
width is less than 981px.