Complete task-02 in homeworks/01
Bottom navigation was styled with position fixed and bottom 0 to
ensure it stays visible at the bottom of the page all the time. It
was also given left 0 and 100% width, so it takes up the entire width
of the page. Using a media query, the bottom navigation was set to be
visible only on the screens with a width of 980px or less.