Style the login container
Added border radius to containers, inputs and a button. Also, hover
and focus effects were added on inputs and the button. Background and
text colors were changed to make the form more visually appealing.
Additionally, the shadow was added to the form to make it stand out
a bit more. The font was changed to Lato because it looks more modern.
Body height and width were adjusted using the calc and min functions
as requested.