Add info container before the start of the quiz
An info container containing a heading, paragraph, and a start button
was added and styled with start.css.
Additionally, global styles providing general styling for all elements
and the body, with colors for both light and dark themes, were added.
In script.js, a click listener was added to the start button, which
calls a function to remove the starting container from the DOM.