Add correct/wrong indicator on answers
When the question is loaded, answer button is disabled because none of
the answers has been selected. When an answer is selected, the answer
button becomes enabled. Once the user submits the answer, all answers
are disabled, along with the answer button. If the answer is correct,
next question button is presented, which leads to the next question
when clicked.
Additionally, updated the loading indicator and added some subtle
effects.